isr80h: Add SYSTEM_COMMAND9_EXIT for process termination
This commit introduces a new system command, SYSTEM_COMMAND9_EXIT, which
allows a process to terminate itself. This functionality is implemented
in isr80h_command9_exit, which calls process_terminate and then switches
to the next task.

Additionally, this commit updates isr80h_register_commands to register
the new command and adds the command enum to isr80h.h.